[ISSUE #9625]♻️Extract metrics catalog inline tests into a sibling module
